+++ src/reversi/reversiview.h
@@ -7,6 +7,8 @@
 #include <optional>
 #include <vector>
 
+class QActionGroup;
+
 class ReversiView : public GameView
 {
     Q_OBJECT
@@ -16,6 +18,8 @@ public:
 
     QList<QAction*> gameActions() override { return m_actions; }
     void activate() override;
+    QByteArray saveState() const override;
+    bool restoreState(const QByteArray& blob) override;
 
 protected:
     void paintEvent(QPaintEvent* event) override;
@@ -45,6 +49,7 @@ private:
 
     QList<QAction*> m_actions;
     QAction* m_undoAction = nullptr;
+    QActionGroup* m_levelGroup = nullptr;
 
     Board m_board;
     Player m_toMove = Player::Black;